+++ CHANGELOG.md
@@ -16,3 +16,4 @@ Most recent version is listed first.
 - logger: https://github.com/komuw/goweb/pull/39
 - logging middleware: https://github.com/komuw/goweb/pull/41
 - quality of life improvements: https://github.com/komuw/goweb/pull/45
+- add unique id generator: https://github.com/komuw/goweb/pull/50